pyproject.toml
[project]
name = "stockspredictionagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=1.5.0",
    "yfinance>=0.2.65",
]
